requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
python3-openid==3.2.0
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.2
opencv-python==4.8.0.74
tensorflow==2.13.0
tensorflow-hub==0.14.0
pytesseract==0.3.10
PyMuPDF==1.23.8
numpy==1.24.3
pdf2image==1.16.3
rembg==2.0.50
